add weather api, house template link

+++ src/main/kotlin/team/mcqueen/smartdisplay/domain/weather/Weather.kt
@@ -2,6 +2,6 @@ package team.mcqueen.smartdisplay.domain.weather
 
 data class Weather(
     var city: String,
-    var temperature: Int,
+    var temperature: Double,
     var weatherType: String
 )
